requirements.txt
langchain>=0.1.0
langchain-community>=0.0.5
langgraph-prebuilt>=0.1.8
openai>=1.0.0
faiss-cpu>=1.7.4
pypdf>=3.13.0
tqdm>=4.64.0
numpy>=1.23.0
pickle-mixin>=1.0.2
python-dotenv>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0